część walidacji

+++ APBD-PJATK-Cw6-s32101/Controllers/AppointmentsController.cs
@@ -25,17 +25,40 @@ public class AppointmentsController(AppointmentService appointmentService) : Con
             return NotFound();
         return Ok(appointmentListDto);
     }
-
+    
+    /*
+       Pacjent musi istnieć i być aktywny.
+       Lekarz musi istnieć i być aktywny.
+       Termin wizyty nie może być w przeszłości.
+       Opis wizyty nie może być pusty i powinien mieć maksymalnie 250 znaków.
+       Lekarz nie może mieć innej zaplanowanej wizyty dokładnie w tym samym terminie.
+     */
     [HttpPost]
     public async Task<IActionResult> CreateAppointmentAsync([FromBody] CreateAppointmentRequestDto obj)
     {
+        if (obj.AppointmentDate < DateTime.UtcNow)
+            return BadRequest("Termin wizyty nie może być w przeszłości");
+        
         await appointmentService.InsertAppointment(obj);
         return Created();
     }
 
+    /*
+       Jeśli wizyta nie istnieje, zwróć 404 Not Found.
+       Pacjent i lekarz muszą istnieć oraz być aktywni.
+       Status musi być jednym z: Scheduled, Completed, Cancelled. - weryfikowane na poziomie bazy
+       Jeśli aktualizowana wizyta ma status Completed, nie pozwól zmienić jej terminu.
+       Przy zmianie terminu sprawdź konflikt z innymi wizytami tego lekarza.
+     */
     [HttpPut("{id:int}")]
     public async Task<IActionResult> UpdateAppointmentAsync(int id, [FromBody] UpdateAppointmentRequestDto obj)
     {
+        // if (obj.AppointmentDate < DateTime.UtcNow) // dopuszczamy aktualizację starych wizyt
+        //     return BadRequest("Termin wizyty nie może być w przeszłości");
+        
+        if ((await appointmentService.GetAppointmentsAsync(id)).FirstOrDefault() == null)
+            return NotFound();
+        
         await appointmentService.UpdateAppointment(id, obj);
         return Ok();
     }
@@ -45,6 +68,13 @@ public class AppointmentsController(AppointmentService appointmentService) : Con
     {
         try
         {
+            var existing = (await appointmentService.GetAppointmentsAsync(id)).FirstOrDefault();
+            if (existing == null)
+                return NotFound();
+
+            if (existing.Status.Equals("Completed", StringComparison.OrdinalIgnoreCase))
+                return Conflict("Nie można usunąć zrealizowanej wizyty");
+            
             await appointmentService.DeleteAppointment(id);
             return NoContent();
         }
